requirements.txt
beeprint
black
complexPyTorch
imageio
ipycanvas
ipython
ipywidgets
kornia
matplotlib
numpy
pandas
Pillow
pprintpp
pyimagej
pytorch_lightning
pyzmq
requests
scipy
six
terminaltables
tifffile
torchviz
wxPython
opencv-python
scikit-image
torchtyping